FULL AAC SUPPORT!!!!!
