+++ src/bit-streams/aac/codebooks.js
@@ -1,1396 +1,1764 @@
-function makeLookupTable(arr) {
-  return arr
-    .map(([idx, len, val])=> [idx, len, ('00000000000000000000000000000000' + val.toString(2)).slice(-len)])
-    .sort((a,b)=>a[1] - b[1])
-    .reduce((arr, p) => {
-      let last = arr[arr.length - 1];
-      if (!last || last.size !== p[1]) {
-        last = {
-          size: p[1],
-        };
-        arr.push(last);
-      }
-      last[p[2]] = p[0];
+export const readCodebookValue = function (codebook, expGolomb) {
+  let bitString = expGolomb.bitReservoir;
 
-      return arr;
-    }, []);
-}
+  for(let i = 0; i < codebook.length; i++) {
+    let o = codebook[i];
+    let subString = bitString.slice(0, o.size);
+    let index = o[subString];
 
-const scaleFactorCB = [
-  [0, 18, 0x3ffe8],
-  [1, 18, 0x3ffe6],
-  [2, 18, 0x3ffe7],
-  [3, 18, 0x3ffe5],
-  [4, 19, 0x7fff5],
-  [5, 19, 0x7fff1],
-  [6, 19, 0x7ffed],
-  [7, 19, 0x7fff6],
-  [8, 19, 0x7ffee],
-  [9, 19, 0x7ffef],
-  [10, 19, 0x7fff0],
-  [11, 19, 0x7fffc],
-  [12, 19, 0x7fffd],
-  [13, 19, 0x7ffff],
-  [14, 19, 0x7fffe],
-  [15, 19, 0x7fff7],
-  [16, 19, 0x7fff8],
-  [17, 19, 0x7fffb],
-  [18, 19, 0x7fff9],
-  [19, 18, 0x3ffe4],
-  [20, 19, 0x7fffa],
-  [21, 18, 0x3ffe3],
-  [22, 17, 0x1ffef],
-  [23, 17, 0x1fff0],
-  [24, 16, 0xfff5],
-  [25, 17, 0x1ffee],
-  [26, 16, 0xfff2],
-  [27, 16, 0xfff3],
-  [28, 16, 0xfff4],
-  [29, 16, 0xfff1],
-  [30, 15, 0x7ff6],
-  [31, 15, 0x7ff7],
-  [32, 14, 0x3ff9],
-  [33, 14, 0x3ff5],
-  [34, 14, 0x3ff7],
-  [35, 14, 0x3ff3],
-  [36, 14, 0x3ff6],
-  [37, 14, 0x3ff2],
-  [38, 13, 0x1ff7],
-  [39, 13, 0x1ff5],
-  [40, 12, 0xff9],
-  [41, 12, 0xff7],
-  [42, 12, 0xff6],
-  [43, 11, 0x7f9],
-  [44, 12, 0xff4],
-  [45, 11, 0x7f8],
-  [46, 10, 0x3f9],
-  [47, 10, 0x3f7],
-  [48, 10, 0x3f5],
-  [49, 9, 0x1f8],
-  [50, 9, 0x1f7],
-  [51, 8, 0xfa],
-  [52, 8, 0xf8],
-  [53, 8, 0xf6],
-  [54, 7, 0x79],
-  [55, 6, 0x3a],
-  [56, 6, 0x38],
-  [57, 5, 0x1a],
-  [58, 4, 0xb],
-  [59, 3, 0x4],
-  [60, 1, 0x0],
-  [61, 4, 0xa],
-  [62, 4, 0xc],
-  [63, 5, 0x1b],
-  [64, 6, 0x39],
-  [65, 6, 0x3b],
-  [66, 7, 0x78],
-  [67, 7, 0x7a],
-  [68, 8, 0xf7],
-  [69, 8, 0xf9],
-  [70, 9, 0x1f6],
-  [71, 9, 0x1f9],
-  [72, 10, 0x3f4],
-  [73, 10, 0x3f6],
-  [74, 10, 0x3f8],
-  [75, 11, 0x7f5],
-  [76, 11, 0x7f4],
-  [77, 11, 0x7f6],
-  [78, 11, 0x7f7],
-  [79, 12, 0xff5],
-  [80, 12, 0xff8],
-  [81, 13, 0x1ff4],
-  [82, 13, 0x1ff6],
-  [83, 13, 0x1ff8],
-  [84, 14, 0x3ff8],
-  [85, 14, 0x3ff4],
-  [86, 16, 0xfff0],
-  [87, 15, 0x7ff4],
-  [88, 16, 0xfff6],
-  [89, 15, 0x7ff5],
-  [90, 18, 0x3ffe2],
-  [91, 19, 0x7ffd9],
-  [92, 19, 0x7ffda],
-  [93, 19, 0x7ffdb],
-  [94, 19, 0x7ffdc],
-  [95, 19, 0x7ffdd],
-  [96, 19, 0x7ffde],
-  [97, 19, 0x7ffd8],
-  [98, 19, 0x7ffd2],
-  [99, 19, 0x7ffd3],
-  [100, 19, 0x7ffd4],
-  [101, 19, 0x7ffd5],
-  [102, 19, 0x7ffd6],
-  [103, 19, 0x7fff2],
-  [104, 19, 0x7ffdf],
-  [105, 19, 0x7ffe7],
-  [106, 19, 0x7ffe8],
-  [107, 19, 0x7ffe9],
-  [108, 19, 0x7ffea],
-  [109, 19, 0x7ffeb],
-  [110, 19, 0x7ffe6],
-  [111, 19, 0x7ffe0],
-  [112, 19, 0x7ffe1],
-  [113, 19, 0x7ffe2],
-  [114, 19, 0x7ffe3],
-  [115, 19, 0x7ffe4],
-  [116, 19, 0x7ffe5],
-  [117, 19, 0x7ffd7],
-  [118, 19, 0x7ffec],
-  [119, 19, 0x7fff4],
-  [120, 19, 0x7fff3]];
+    if (typeof index === 'number') {
+      expGolomb.readBits(o.size);
+      return index;
+    }
+  }
+  return -999999;
+  //expGolomb.bitStream = '';
+};
 
-const spectrumCB = [
-[], // ZERO_HCB
-[
-  [0, 11, 0x7f8],
-  [1, 9, 0x1f1],
-  [2, 11, 0x7fd],
-  [3, 10, 0x3f5],
-  [4, 7, 0x68],
-  [5, 10, 0x3f0],
-  [6, 11, 0x7f7],
-  [7, 9, 0x1ec],
-  [8, 11, 0x7f5],
-  [9, 10, 0x3f1],
-  [10, 7, 0x72],
-  [11, 10, 0x3f4],
-  [12, 7, 0x74],
-  [13, 5, 0x11],
-  [14, 7, 0x76],
-  [15, 9, 0x1eb],
-  [16, 7, 0x6c],
-  [17, 10, 0x3f6],
-  [18, 11, 0x7fc],
-  [19, 9, 0x1e1],
-  [20, 11, 0x7f1],
-  [21, 9, 0x1f0],
-  [22, 7, 0x61],
-  [23, 9, 0x1f6],
-  [24, 11, 0x7f2],
-  [25, 9, 0x1ea],
-  [26, 11, 0x7fb],
-  [27, 9, 0x1f2],
-  [28, 7, 0x69],
-  [29, 9, 0x1ed],
-  [30, 7, 0x77],
-  [31, 5, 0x17],
-  [32, 7, 0x6f],
-  [33, 9, 0x1e6],
-  [34, 7, 0x64],
-  [35, 9, 0x1e5],
-  [36, 7, 0x67],
-  [37, 5, 0x15],
-  [38, 7, 0x62],
-  [39, 5, 0x12],
-  [40, 1, 0x0],
-  [41, 5, 0x14],
-  [42, 7, 0x65],
-  [43, 5, 0x16],
-  [44, 7, 0x6d],
-  [45, 9, 0x1e9],
-  [46, 7, 0x63],
-  [47, 9, 0x1e4],
-  [48, 7, 0x6b],
-  [49, 5, 0x13],
-  [50, 7, 0x71],
-  [51, 9, 0x1e3],
-  [52, 7, 0x70],
-  [53, 9, 0x1f3],
-  [54, 11, 0x7fe],
-  [55, 9, 0x1e7],
-  [56, 11, 0x7f3],
-  [57, 9, 0x1ef],
-  [58, 7, 0x60],
-  [59, 9, 0x1ee],
-  [60, 11, 0x7f0],
-  [61, 9, 0x1e2],
-  [62, 11, 0x7fa],
-  [63, 10, 0x3f3],
-  [64, 7, 0x6a],
-  [65, 9, 0x1e8],
-  [66, 7, 0x75],
-  [67, 5, 0x10],
-  [68, 7, 0x73],
-  [69, 9, 0x1f4],
-  [70, 7, 0x6e],
-  [71, 10, 0x3f7],
-  [72, 11, 0x7f6],
-  [73, 9, 0x1e0],
-  [74, 11, 0x7f9],
-  [75, 10, 0x3f2],
-  [76, 7, 0x66],
-  [77, 9, 0x1f5],
-  [78, 11, 0x7ff],
-  [79, 9, 0x1f7],
-  [80, 11, 0x7f4]
-], [
-  [0, 9, 0x1f3],
-  [1, 7, 0x6f],
-  [2, 9, 0x1fd],
-  [3, 8, 0xeb],
-  [4, 6, 0x23],
-  [5, 8, 0xea],
-  [6, 9, 0x1f7],
-  [7, 8, 0xe8],
-  [8, 9, 0x1fa],
-  [9, 8, 0xf2],
-  [10, 6, 0x2d],
-  [11, 7, 0x70],
-  [12, 6, 0x20],
-  [13, 5, 0x6],
-  [14, 6, 0x2b],
-  [15, 7, 0x6e],
-  [16, 6, 0x28],
-  [17, 8, 0xe9],
-  [18, 9, 0x1f9],
-  [19, 7, 0x66],
-  [20, 8, 0xf8],
-  [21, 8, 0xe7],
-  [22, 6, 0x1b],
-  [23, 8, 0xf1],
-  [24, 9, 0x1f4],
-  [25, 7, 0x6b],
-  [26, 9, 0x1f5],
-  [27, 8, 0xec],
-  [28, 6, 0x2a],
-  [29, 7, 0x6c],
-  [30, 6, 0x2c],
-  [31, 5, 0xa],
-  [32, 6, 0x27],
-  [33, 7, 0x67],
-  [34, 6, 0x1a],
-  [35, 8, 0xf5],
-  [36, 6, 0x24],
-  [37, 5, 0x8],
-  [38, 6, 0x1f],
-  [39, 5, 0x9],
-  [40, 3, 0x0],
-  [41, 5, 0x7],
-  [42, 6, 0x1d],
-  [43, 5, 0xb],
-  [44, 6, 0x30],
-  [45, 8, 0xef],
-  [46, 6, 0x1c],
-  [47, 7, 0x64],
-  [48, 6, 0x1e],
-  [49, 5, 0xc],
-  [50, 6, 0x29],
-  [51, 8, 0xf3],
-  [52, 6, 0x2f],
-  [53, 8, 0xf0],
-  [54, 9, 0x1fc],
-  [55, 7, 0x71],
-  [56, 9, 0x1f2],
-  [57, 8, 0xf4],
-  [58, 6, 0x21],
-  [59, 8, 0xe6],
-  [60, 8, 0xf7],
-  [61, 7, 0x68],
-  [62, 9, 0x1f8],
-  [63, 8, 0xee],
-  [64, 6, 0x22],
-  [65, 7, 0x65],
-  [66, 6, 0x31],
-  [67, 4, 0x2],
-  [68, 6, 0x26],
-  [69, 8, 0xed],
-  [70, 6, 0x25],
-  [71, 7, 0x6a],
-  [72, 9, 0x1fb],
-  [73, 7, 0x72],
-  [74, 9, 0x1fe],
-  [75, 7, 0x69],
-  [76, 6, 0x2e],
-  [77, 8, 0xf6],
-  [78, 9, 0x1ff],
-  [79, 7, 0x6d],
-  [80, 9, 0x1f6]
-], [
-  [0, 1, 0x0],
-  [1, 4, 0x9],
-  [2, 8, 0xef],
-  [3, 4, 0xb],
-  [4, 5, 0x19],
-  [5, 8, 0xf0],
-  [6, 9, 0x1eb],
-  [7, 9, 0x1e6],
-  [8, 10, 0x3f2],
-  [9, 4, 0xa],
-  [10, 6, 0x35],
-  [11, 9, 0x1ef],
-  [12, 6, 0x34],
-  [13, 6, 0x37],
-  [14, 9, 0x1e9],
-  [15, 9, 0x1ed],
-  [16, 9, 0x1e7],
-  [17, 10, 0x3f3],
-  [18, 9, 0x1ee],
-  [19, 10, 0x3ed],
-  [20, 13, 0x1ffa],
-  [21, 9, 0x1ec],
-  [22, 9, 0x1f2],
-  [23, 11, 0x7f9],
-  [24, 11, 0x7f8],
-  [25, 10, 0x3f8],
-  [26, 12, 0xff8],
-  [27, 4, 0x8],
-  [28, 6, 0x38],
-  [29, 10, 0x3f6],
-  [30, 6, 0x36],
-  [31, 7, 0x75],
-  [32, 10, 0x3f1],
-  [33, 10, 0x3eb],
-  [34, 10, 0x3ec],
-  [35, 12, 0xff4],
-  [36, 5, 0x18],
-  [37, 7, 0x76],
-  [38, 11, 0x7f4],
-  [39, 6, 0x39],
-  [40, 7, 0x74],
-  [41, 10, 0x3ef],
-  [42, 9, 0x1f3],
-  [43, 9, 0x1f4],
-  [44, 11, 0x7f6],
-  [45, 9, 0x1e8],
-  [46, 10, 0x3ea],
-  [47, 13, 0x1ffc],
-  [48, 8, 0xf2],
-  [49, 9, 0x1f1],
-  [50, 12, 0xffb],
-  [51, 10, 0x3f5],
-  [52, 11, 0x7f3],
-  [53, 12, 0xffc],
-  [54, 8, 0xee],
-  [55, 10, 0x3f7],
-  [56, 15, 0x7ffe],
-  [57, 9, 0x1f0],
-  [58, 11, 0x7f5],
-  [59, 15, 0x7ffd],
-  [60, 13, 0x1ffb],
-  [61, 14, 0x3ffa],
-  [62, 16, 0xffff],
-  [63, 8, 0xf1],
-  [64, 10, 0x3f0],
-  [65, 14, 0x3ffc],
-  [66, 9, 0x1ea],
-  [67, 10, 0x3ee],
-  [68, 14, 0x3ffb],
-  [69, 12, 0xff6],
-  [70, 12, 0xffa],
-  [71, 15, 0x7ffc],
-  [72, 11, 0x7f2],
-  [73, 12, 0xff5],
-  [74, 16, 0xfffe],
-  [75, 10, 0x3f4],
-  [76, 11, 0x7f7],
-  [77, 15, 0x7ffb],
-  [78, 12, 0xff7],
-  [79, 12, 0xff9],
-  [80, 15, 0x7ffa]
-], [
-  [0, 4, 0x7],
-  [1, 5, 0x16],
-  [2, 8, 0xf6],
-  [3, 5, 0x18],
-  [4, 4, 0x8],
-  [5, 8, 0xef],
-  [6, 9, 0x1ef],
-  [7, 8, 0xf3],
-  [8, 11, 0x7f8],
-  [9, 5, 0x19],
-  [10, 5, 0x17],
-  [11, 8, 0xed],
-  [12, 5, 0x15],
-  [13, 4, 0x1],
-  [14, 8, 0xe2],
-  [15, 8, 0xf0],
-  [16, 7, 0x70],
-  [17, 10, 0x3f0],
-  [18, 9, 0x1ee],
-  [19, 8, 0xf1],
-  [20, 11, 0x7fa],
-  [21, 8, 0xee],
-  [22, 8, 0xe4],
-  [23, 10, 0x3f2],
-  [24, 11, 0x7f6],
-  [25, 10, 0x3ef],
-  [26, 11, 0x7fd],
-  [27, 4, 0x5],
-  [28, 5, 0x14],
-  [29, 8, 0xf2],
-  [30, 4, 0x9],
-  [31, 4, 0x4],
-  [32, 8, 0xe5],
-  [33, 8, 0xf4],
-  [34, 8, 0xe8],
-  [35, 10, 0x3f4],
-  [36, 4, 0x6],
-  [37, 4, 0x2],
-  [38, 8, 0xe7],
-  [39, 4, 0x3],
-  [40, 4, 0x0],
-  [41, 7, 0x6b],
-  [42, 8, 0xe3],
-  [43, 7, 0x69],
-  [44, 9, 0x1f3],
-  [45, 8, 0xeb],
-  [46, 8, 0xe6],
-  [47, 10, 0x3f6],
-  [48, 7, 0x6e],
-  [49, 7, 0x6a],
-  [50, 9, 0x1f4],
-  [51, 10, 0x3ec],
-  [52, 9, 0x1f0],
-  [53, 10, 0x3f9],
-  [54, 8, 0xf5],
-  [55, 8, 0xec],
-  [56, 11, 0x7fb],
-  [57, 8, 0xea],
-  [58, 7, 0x6f],
-  [59, 10, 0x3f7],
-  [60, 11, 0x7f9],
-  [61, 10, 0x3f3],
-  [62, 12, 0xfff],
-  [63, 8, 0xe9],
-  [64, 7, 0x6d],
-  [65, 10, 0x3f8],
-  [66, 7, 0x6c],
-  [67, 7, 0x68],
-  [68, 9, 0x1f5],
-  [69, 10, 0x3ee],
-  [70, 9, 0x1f2],
-  [71, 11, 0x7f4],
-  [72, 11, 0x7f7],
-  [73, 10, 0x3f1],
-  [74, 12, 0xffe],
-  [75, 10, 0x3ed],
-  [76, 9, 0x1f1],
-  [77, 11, 0x7f5],
-  [78, 11, 0x7fe],
-  [79, 10, 0x3f5],
-  [80, 11, 0x7fc]
-], [
-  [0, 13, 0x1fff],
-  [1, 12, 0xff7],
-  [2, 11, 0x7f4],
-  [3, 11, 0x7e8],
-  [4, 10, 0x3f1],
-  [5, 11, 0x7ee],
-  [6, 11, 0x7f9],
-  [7, 12, 0xff8],
-  [8, 13, 0x1ffd],
-  [9, 12, 0xffd],
-  [10, 11, 0x7f1],
-  [11, 10, 0x3e8],
-  [12, 9, 0x1e8],
-  [13, 8, 0xf0],
-  [14, 9, 0x1ec],
-  [15, 10, 0x3ee],
-  [16, 11, 0x7f2],
-  [17, 12, 0xffa],
-  [18, 12, 0xff4],
-  [19, 10, 0x3ef],
-  [20, 9, 0x1f2],
-  [21, 8, 0xe8],
-  [22, 7, 0x70],
-  [23, 8, 0xec],
-  [24, 9, 0x1f0],
-  [25, 10, 0x3ea],
-  [26, 11, 0x7f3],
-  [27, 11, 0x7eb],
-  [28, 9, 0x1eb],
-  [29, 8, 0xea],
-  [30, 5, 0x1a],
-  [31, 4, 0x8],
-  [32, 5, 0x19],
-  [33, 8, 0xee],
-  [34, 9, 0x1ef],
-  [35, 11, 0x7ed],
-  [36, 10, 0x3f0],
-  [37, 8, 0xf2],
-  [38, 7, 0x73],
-  [39, 4, 0xb],
-  [40, 1, 0x0],
-  [41, 4, 0xa],
-  [42, 7, 0x71],
-  [43, 8, 0xf3],
-  [44, 11, 0x7e9],
-  [45, 11, 0x7ef],
-  [46, 9, 0x1ee],
-  [47, 8, 0xef],
-  [48, 5, 0x18],
-  [49, 4, 0x9],
-  [50, 5, 0x1b],
-  [51, 8, 0xeb],
-  [52, 9, 0x1e9],
-  [53, 11, 0x7ec],
-  [54, 11, 0x7f6],
-  [55, 10, 0x3eb],
-  [56, 9, 0x1f3],
-  [57, 8, 0xed],
-  [58, 7, 0x72],
-  [59, 8, 0xe9],
-  [60, 9, 0x1f1],
-  [61, 10, 0x3ed],
-  [62, 11, 0x7f7],
-  [63, 12, 0xff6],
-  [64, 11, 0x7f0],
-  [65, 10, 0x3e9],
-  [66, 9, 0x1ed],
-  [67, 8, 0xf1],
-  [68, 9, 0x1ea],
-  [69, 10, 0x3ec],
-  [70, 11, 0x7f8],
-  [71, 12, 0xff9],
-  [72, 13, 0x1ffc],
-  [73, 12, 0xffc],
-  [74, 12, 0xff5],
-  [75, 11, 0x7ea],
-  [76, 10, 0x3f3],
-  [77, 10, 0x3f2],
-  [78, 11, 0x7f5],
-  [79, 12, 0xffb],
-  [80, 13, 0x1ffe]
-], [
-  [0, 11, 0x7fe],
-  [1, 10, 0x3fd],
-  [2, 9, 0x1f1],
-  [3, 9, 0x1eb],
-  [4, 9, 0x1f4],
-  [5, 9, 0x1ea],
-  [6, 9, 0x1f0],
-  [7, 10, 0x3fc],
-  [8, 11, 0x7fd],
-  [9, 10, 0x3f6],
-  [10, 9, 0x1e5],
-  [11, 8, 0xea],
-  [12, 7, 0x6c],
-  [13, 7, 0x71],
-  [14, 7, 0x68],
-  [15, 8, 0xf0],
-  [16, 9, 0x1e6],
-  [17, 10, 0x3f7],
-  [18, 9, 0x1f3],
-  [19, 8, 0xef],
-  [20, 6, 0x32],
-  [21, 6, 0x27],
-  [22, 6, 0x28],
-  [23, 6, 0x26],
-  [24, 6, 0x31],
-  [25, 8, 0xeb],
-  [26, 9, 0x1f7],
-  [27, 9, 0x1e8],
-  [28, 7, 0x6f],
-  [29, 6, 0x2e],
-  [30, 4, 0x8],
-  [31, 4, 0x4],
-  [32, 4, 0x6],
-  [33, 6, 0x29],
-  [34, 7, 0x6b],
-  [35, 9, 0x1ee],
-  [36, 9, 0x1ef],
-  [37, 7, 0x72],
-  [38, 6, 0x2d],
-  [39, 4, 0x2],
-  [40, 4, 0x0],
-  [41, 4, 0x3],
-  [42, 6, 0x2f],
-  [43, 7, 0x73],
-  [44, 9, 0x1fa],
-  [45, 9, 0x1e7],
-  [46, 7, 0x6e],
-  [47, 6, 0x2b],
-  [48, 4, 0x7],
-  [49, 4, 0x1],
-  [50, 4, 0x5],
-  [51, 6, 0x2c],
-  [52, 7, 0x6d],
-  [53, 9, 0x1ec],
-  [54, 9, 0x1f9],
-  [55, 8, 0xee],
-  [56, 6, 0x30],
-  [57, 6, 0x24],
-  [58, 6, 0x2a],
-  [59, 6, 0x25],
-  [60, 6, 0x33],
-  [61, 8, 0xec],
-  [62, 9, 0x1f2],
-  [63, 10, 0x3f8],
-  [64, 9, 0x1e4],
-  [65, 8, 0xed],
-  [66, 7, 0x6a],
-  [67, 7, 0x70],
-  [68, 7, 0x69],
-  [69, 7, 0x74],
-  [70, 8, 0xf1],
-  [71, 10, 0x3fa],
-  [72, 11, 0x7ff],
-  [73, 10, 0x3f9],
-  [74, 9, 0x1f6],
-  [75, 9, 0x1ed],
-  [76, 9, 0x1f8],
-  [77, 9, 0x1e9],
-  [78, 9, 0x1f5],
-  [79, 10, 0x3fb],
-  [80, 11, 0x7fc]
-], [
-  [0, 1, 0x0],
-  [1, 3, 0x5],
-  [2, 6, 0x37],
-  [3, 7, 0x74],
-  [4, 8, 0xf2],
-  [5, 9, 0x1eb],
-  [6, 10, 0x3ed],
-  [7, 11, 0x7f7],
-  [8, 3, 0x4],
-  [9, 4, 0xc],
-  [10, 6, 0x35],
-  [11, 7, 0x71],
-  [12, 8, 0xec],
-  [13, 8, 0xee],
-  [14, 9, 0x1ee],
-  [15, 9, 0x1f5],
-  [16, 6, 0x36],
-  [17, 6, 0x34],
-  [18, 7, 0x72],
-  [19, 8, 0xea],
-  [20, 8, 0xf1],
-  [21, 9, 0x1e9],
-  [22, 9, 0x1f3],
-  [23, 10, 0x3f5],
-  [24, 7, 0x73],
-  [25, 7, 0x70],
-  [26, 8, 0xeb],
-  [27, 8, 0xf0],
-  [28, 9, 0x1f1],
-  [29, 9, 0x1f0],
-  [30, 10, 0x3ec],
-  [31, 10, 0x3fa],
-  [32, 8, 0xf3],
-  [33, 8, 0xed],
-  [34, 9, 0x1e8],
-  [35, 9, 0x1ef],
-  [36, 10, 0x3ef],
-  [37, 10, 0x3f1],
-  [38, 10, 0x3f9],
-  [39, 11, 0x7fb],
-  [40, 9, 0x1ed],
-  [41, 8, 0xef],
-  [42, 9, 0x1ea],
-  [43, 9, 0x1f2],
-  [44, 10, 0x3f3],
-  [45, 10, 0x3f8],
-  [46, 11, 0x7f9],
-  [47, 11, 0x7fc],
-  [48, 10, 0x3ee],
-  [49, 9, 0x1ec],
-  [50, 9, 0x1f4],
-  [51, 10, 0x3f4],
-  [52, 10, 0x3f7],
-  [53, 11, 0x7f8],
-  [54, 12, 0xffd],
-  [55, 12, 0xffe],
-  [56, 11, 0x7f6],
-  [57, 10, 0x3f0],
-  [58, 10, 0x3f2],
-  [59, 10, 0x3f6],
-  [60, 11, 0x7fa],
-  [61, 11, 0x7fd],
-  [62, 12, 0xffc],
-  [63, 12, 0xfff]
-], [
-  [0, 5, 0xe],
-  [1, 4, 0x5],
-  [2, 5, 0x10],
-  [3, 6, 0x30],
-  [4, 7, 0x6f],
-  [5, 8, 0xf1],
-  [6, 9, 0x1fa],
-  [7, 10, 0x3fe],
-  [8, 4, 0x3],
-  [9, 3, 0x0],
-  [10, 4, 0x4],
-  [11, 5, 0x12],
-  [12, 6, 0x2c],
-  [13, 7, 0x6a],
-  [14, 7, 0x75],
-  [15, 8, 0xf8],
-  [16, 5, 0xf],
-  [17, 4, 0x2],
-  [18, 4, 0x6],
-  [19, 5, 0x14],
-  [20, 6, 0x2e],
-  [21, 7, 0x69],
-  [22, 7, 0x72],
-  [23, 8, 0xf5],
-  [24, 6, 0x2f],
-  [25, 5, 0x11],
-  [26, 5, 0x13],
-  [27, 6, 0x2a],
-  [28, 6, 0x32],
-  [29, 7, 0x6c],
-  [30, 8, 0xec],
-  [31, 8, 0xfa],
-  [32, 7, 0x71],
-  [33, 6, 0x2b],
-  [34, 6, 0x2d],
-  [35, 6, 0x31],
-  [36, 7, 0x6d],
-  [37, 7, 0x70],
-  [38, 8, 0xf2],
-  [39, 9, 0x1f9],
-  [40, 8, 0xef],
-  [41, 7, 0x68],
-  [42, 6, 0x33],
-  [43, 7, 0x6b],
-  [44, 7, 0x6e],
-  [45, 8, 0xee],
-  [46, 8, 0xf9],
-  [47, 10, 0x3fc],
-  [48, 9, 0x1f8],
-  [49, 7, 0x74],
-  [50, 7, 0x73],
-  [51, 8, 0xed],
-  [52, 8, 0xf0],
-  [53, 8, 0xf6],
-  [54, 9, 0x1f6],
-  [55, 9, 0x1fd],
-  [56, 10, 0x3fd],
-  [57, 8, 0xf3],
-  [58, 8, 0xf4],
-  [59, 8, 0xf7],
-  [60, 9, 0x1f7],
-  [61, 9, 0x1fb],
-  [62, 9, 0x1fc],
-  [63, 10, 0x3ff]
-], [
-  [0, 1, 0x0],
-  [1, 3, 0x5],
-  [2, 6, 0x37],
-  [3, 8, 0xe7],
-  [4, 9, 0x1de],
-  [5, 10, 0x3ce],
-  [6, 10, 0x3d9],
-  [7, 11, 0x7c8],
-  [8, 11, 0x7cd],
-  [9, 12, 0xfc8],
-  [10, 12, 0xfdd],
-  [11, 13, 0x1fe4],
-  [12, 13, 0x1fec],
-  [13, 3, 0x4],
-  [14, 4, 0xc],
-  [15, 6, 0x35],
-  [16, 7, 0x72],
-  [17, 8, 0xea],
-  [18, 8, 0xed],
-  [19, 9, 0x1e2],
-  [20, 10, 0x3d1],
-  [21, 10, 0x3d3],
-  [22, 10, 0x3e0],
-  [23, 11, 0x7d8],
-  [24, 12, 0xfcf],
-  [25, 12, 0xfd5],
-  [26, 6, 0x36],
-  [27, 6, 0x34],
-  [28, 7, 0x71],
-  [29, 8, 0xe8],
-  [30, 8, 0xec],
-  [31, 9, 0x1e1],
-  [32, 10, 0x3cf],
-  [33, 10, 0x3dd],
-  [34, 10, 0x3db],
-  [35, 11, 0x7d0],
-  [36, 12, 0xfc7],
-  [37, 12, 0xfd4],
-  [38, 12, 0xfe4],
-  [39, 8, 0xe6],
-  [40, 7, 0x70],
-  [41, 8, 0xe9],
-  [42, 9, 0x1dd],
-  [43, 9, 0x1e3],
-  [44, 10, 0x3d2],
-  [45, 10, 0x3dc],
-  [46, 11, 0x7cc],
-  [47, 11, 0x7ca],
-  [48, 11, 0x7de],
-  [49, 12, 0xfd8],
-  [50, 12, 0xfea],
-  [51, 13, 0x1fdb],
-  [52, 9, 0x1df],
-  [53, 8, 0xeb],
-  [54, 9, 0x1dc],
-  [55, 9, 0x1e6],
-  [56, 10, 0x3d5],
-  [57, 10, 0x3de],
-  [58, 11, 0x7cb],
-  [59, 11, 0x7dd],
-  [60, 11, 0x7dc],
-  [61, 12, 0xfcd],
-  [62, 12, 0xfe2],
-  [63, 12, 0xfe7],
-  [64, 13, 0x1fe1],
-  [65, 10, 0x3d0],
-  [66, 9, 0x1e0],
-  [67, 9, 0x1e4],
-  [68, 10, 0x3d6],
-  [69, 11, 0x7c5],
-  [70, 11, 0x7d1],
-  [71, 11, 0x7db],
-  [72, 12, 0xfd2],
-  [73, 11, 0x7e0],
-  [74, 12, 0xfd9],
-  [75, 12, 0xfeb],
-  [76, 13, 0x1fe3],
-  [77, 13, 0x1fe9],
-  [78, 11, 0x7c4],
-  [79, 9, 0x1e5],
-  [80, 10, 0x3d7],
-  [81, 11, 0x7c6],
-  [82, 11, 0x7cf],
-  [83, 11, 0x7da],
-  [84, 12, 0xfcb],
-  [85, 12, 0xfda],
-  [86, 12, 0xfe3],
-  [87, 12, 0xfe9],
-  [88, 13, 0x1fe6],
-  [89, 13, 0x1ff3],
-  [90, 13, 0x1ff7],
-  [91, 11, 0x7d3],
-  [92, 10, 0x3d8],
-  [93, 10, 0x3e1],
-  [94, 11, 0x7d4],
-  [95, 11, 0x7d9],
-  [96, 12, 0xfd3],
-  [97, 12, 0xfde],
-  [98, 13, 0x1fdd],
-  [99, 13, 0x1fd9],
-  [100, 13, 0x1fe2],
-  [101, 13, 0x1fea],
-  [102, 13, 0x1ff1],
-  [103, 13, 0x1ff6],
-  [104, 11, 0x7d2],
-  [105, 10, 0x3d4],
-  [106, 10, 0x3da],
-  [107, 11, 0x7c7],
-  [108, 11, 0x7d7],
-  [109, 11, 0x7e2],
-  [110, 12, 0xfce],
-  [111, 12, 0xfdb],
-  [112, 13, 0x1fd8],
-  [113, 13, 0x1fee],
-  [114, 14, 0x3ff0],
-  [115, 13, 0x1ff4],
-  [116, 14, 0x3ff2],
-  [117, 11, 0x7e1],
-  [118, 10, 0x3df],
-  [119, 11, 0x7c9],
-  [120, 11, 0x7d6],
-  [121, 12, 0xfca],
-  [122, 12, 0xfd0],
-  [123, 12, 0xfe5],
-  [124, 12, 0xfe6],
-  [125, 13, 0x1feb],
-  [126, 13, 0x1fef],
-  [127, 14, 0x3ff3],
-  [128, 14, 0x3ff4],
-  [129, 14, 0x3ff5],
-  [130, 12, 0xfe0],
-  [131, 11, 0x7ce],
-  [132, 11, 0x7d5],
-  [133, 12, 0xfc6],
-  [134, 12, 0xfd1],
-  [135, 12, 0xfe1],
-  [136, 13, 0x1fe0],
-  [137, 13, 0x1fe8],
-  [138, 13, 0x1ff0],
-  [139, 14, 0x3ff1],
-  [140, 14, 0x3ff8],
-  [141, 14, 0x3ff6],
-  [142, 15, 0x7ffc],
-  [143, 12, 0xfe8],
-  [144, 11, 0x7df],
-  [145, 12, 0xfc9],
-  [146, 12, 0xfd7],
-  [147, 12, 0xfdc],
-  [148, 13, 0x1fdc],
-  [149, 13, 0x1fdf],
-  [150, 13, 0x1fed],
-  [151, 13, 0x1ff5],
-  [152, 14, 0x3ff9],
-  [153, 14, 0x3ffb],
-  [154, 15, 0x7ffd],
-  [155, 15, 0x7ffe],
-  [156, 13, 0x1fe7],
-  [157, 12, 0xfcc],
-  [158, 12, 0xfd6],
-  [159, 12, 0xfdf],
-  [160, 13, 0x1fde],
-  [161, 13, 0x1fda],
-  [162, 13, 0x1fe5],
-  [163, 13, 0x1ff2],
-  [164, 14, 0x3ffa],
-  [165, 14, 0x3ff7],
-  [166, 14, 0x3ffc],
-  [167, 14, 0x3ffd],
-  [168, 15, 0x7fff]
-], [
-  [0, 6, 0x22],
-  [1, 5, 0x8],
-  [2, 6, 0x1d],
-  [3, 6, 0x26],
-  [4, 7, 0x5f],
-  [5, 8, 0xd3],
-  [6, 9, 0x1cf],
-  [7, 10, 0x3d0],
-  [8, 10, 0x3d7],
-  [9, 10, 0x3ed],
-  [10, 11, 0x7f0],
-  [11, 11, 0x7f6],
-  [12, 12, 0xffd],
-  [13, 5, 0x7],
-  [14, 4, 0x0],
-  [15, 4, 0x1],
-  [16, 5, 0x9],
-  [17, 6, 0x20],
-  [18, 7, 0x54],
-  [19, 7, 0x60],
-  [20, 8, 0xd5],
-  [21, 8, 0xdc],
-  [22, 9, 0x1d4],
-  [23, 10, 0x3cd],
-  [24, 10, 0x3de],
-  [25, 11, 0x7e7],
-  [26, 6, 0x1c],
-  [27, 4, 0x2],
-  [28, 5, 0x6],
-  [29, 5, 0xc],
-  [30, 6, 0x1e],
-  [31, 6, 0x28],
-  [32, 7, 0x5b],
-  [33, 8, 0xcd],
-  [34, 8, 0xd9],
-  [35, 9, 0x1ce],
-  [36, 9, 0x1dc],
-  [37, 10, 0x3d9],
-  [38, 10, 0x3f1],
-  [39, 6, 0x25],
-  [40, 5, 0xb],
-  [41, 5, 0xa],
-  [42, 5, 0xd],
-  [43, 6, 0x24],
-  [44, 7, 0x57],
-  [45, 7, 0x61],
-  [46, 8, 0xcc],
-  [47, 8, 0xdd],
-  [48, 9, 0x1cc],
-  [49, 9, 0x1de],
-  [50, 10, 0x3d3],
-  [51, 10, 0x3e7],
-  [52, 7, 0x5d],
-  [53, 6, 0x21],
-  [54, 6, 0x1f],
-  [55, 6, 0x23],
-  [56, 6, 0x27],
-  [57, 7, 0x59],
-  [58, 7, 0x64],
-  [59, 8, 0xd8],
-  [60, 8, 0xdf],
-  [61, 9, 0x1d2],
-  [62, 9, 0x1e2],
-  [63, 10, 0x3dd],
-  [64, 10, 0x3ee],
-  [65, 8, 0xd1],
-  [66, 7, 0x55],
-  [67, 6, 0x29],
-  [68, 7, 0x56],
-  [69, 7, 0x58],
-  [70, 7, 0x62],
-  [71, 8, 0xce],
-  [72, 8, 0xe0],
-  [73, 8, 0xe2],
-  [74, 9, 0x1da],
-  [75, 10, 0x3d4],
-  [76, 10, 0x3e3],
-  [77, 11, 0x7eb],
-  [78, 9, 0x1c9],
-  [79, 7, 0x5e],
-  [80, 7, 0x5a],
-  [81, 7, 0x5c],
-  [82, 7, 0x63],
-  [83, 8, 0xca],
-  [84, 8, 0xda],
-  [85, 9, 0x1c7],
-  [86, 9, 0x1ca],
-  [87, 9, 0x1e0],
-  [88, 10, 0x3db],
-  [89, 10, 0x3e8],
-  [90, 11, 0x7ec],
-  [91, 9, 0x1e3],
-  [92, 8, 0xd2],
-  [93, 8, 0xcb],
-  [94, 8, 0xd0],
-  [95, 8, 0xd7],
-  [96, 8, 0xdb],
-  [97, 9, 0x1c6],
-  [98, 9, 0x1d5],
-  [99, 9, 0x1d8],
-  [100, 10, 0x3ca],
-  [101, 10, 0x3da],
-  [102, 11, 0x7ea],
-  [103, 11, 0x7f1],
-  [104, 9, 0x1e1],
-  [105, 8, 0xd4],
-  [106, 8, 0xcf],
-  [107, 8, 0xd6],
-  [108, 8, 0xde],
-  [109, 8, 0xe1],
-  [110, 9, 0x1d0],
-  [111, 9, 0x1d6],
-  [112, 10, 0x3d1],
-  [113, 10, 0x3d5],
-  [114, 10, 0x3f2],
-  [115, 11, 0x7ee],
-  [116, 11, 0x7fb],
-  [117, 10, 0x3e9],
-  [118, 9, 0x1cd],
-  [119, 9, 0x1c8],
-  [120, 9, 0x1cb],
-  [121, 9, 0x1d1],
-  [122, 9, 0x1d7],
-  [123, 9, 0x1df],
-  [124, 10, 0x3cf],
-  [125, 10, 0x3e0],
-  [126, 10, 0x3ef],
-  [127, 11, 0x7e6],
-  [128, 11, 0x7f8],
-  [129, 12, 0xffa],
-  [130, 10, 0x3eb],
-  [131, 9, 0x1dd],
-  [132, 9, 0x1d3],
-  [133, 9, 0x1d9],
-  [134, 9, 0x1db],
-  [135, 10, 0x3d2],
-  [136, 10, 0x3cc],
-  [137, 10, 0x3dc],
-  [138, 10, 0x3ea],
-  [139, 11, 0x7ed],
-  [140, 11, 0x7f3],
-  [141, 11, 0x7f9],
-  [142, 12, 0xff9],
-  [143, 11, 0x7f2],
-  [144, 10, 0x3ce],
-  [145, 9, 0x1e4],
-  [146, 10, 0x3cb],
-  [147, 10, 0x3d8],
-  [148, 10, 0x3d6],
-  [149, 10, 0x3e2],
-  [150, 10, 0x3e5],
-  [151, 11, 0x7e8],
-  [152, 11, 0x7f4],
-  [153, 11, 0x7f5],
-  [154, 11, 0x7f7],
-  [155, 12, 0xffb],
-  [156, 11, 0x7fa],
-  [157, 10, 0x3ec],
-  [158, 10, 0x3df],
-  [159, 10, 0x3e1],
-  [160, 10, 0x3e4],
-  [161, 10, 0x3e6],
-  [162, 10, 0x3f0],
-  [163, 11, 0x7e9],
-  [164, 11, 0x7ef],
-  [165, 12, 0xff8],
-  [166, 12, 0xffe],
-  [167, 12, 0xffc],
-  [168, 12, 0xfff]
-], [
-  [0, 4, 0x0],
-  [1, 5, 0x6],
-  [2, 6, 0x19],
-  [3, 7, 0x3d],
-  [4, 8, 0x9c],
-  [5, 8, 0xc6],
-  [6, 9, 0x1a7],
-  [7, 10, 0x390],
-  [8, 10, 0x3c2],
-  [9, 10, 0x3df],
-  [10, 11, 0x7e6],
-  [11, 11, 0x7f3],
-  [12, 12, 0xffb],
-  [13, 11, 0x7ec],
-  [14, 12, 0xffa],
-  [15, 12, 0xffe],
-  [16, 10, 0x38e],
-  [17, 5, 0x5],
-  [18, 4, 0x1],
-  [19, 5, 0x8],
-  [20, 6, 0x14],
-  [21, 7, 0x37],
-  [22, 7, 0x42],
-  [23, 8, 0x92],
-  [24, 8, 0xaf],
-  [25, 9, 0x191],
-  [26, 9, 0x1a5],
-  [27, 9, 0x1b5],
-  [28, 10, 0x39e],
-  [29, 10, 0x3c0],
-  [30, 10, 0x3a2],
-  [31, 10, 0x3cd],
-  [32, 11, 0x7d6],
-  [33, 8, 0xae],
-  [34, 6, 0x17],
-  [35, 5, 0x7],
-  [36, 5, 0x9],
-  [37, 6, 0x18],
-  [38, 7, 0x39],
-  [39, 7, 0x40],
-  [40, 8, 0x8e],
-  [41, 8, 0xa3],
-  [42, 8, 0xb8],
-  [43, 9, 0x199],
-  [44, 9, 0x1ac],
-  [45, 9, 0x1c1],
-  [46, 10, 0x3b1],
-  [47, 10, 0x396],
-  [48, 10, 0x3be],
-  [49, 10, 0x3ca],
-  [50, 8, 0x9d],
-  [51, 7, 0x3c],
-  [52, 6, 0x15],
-  [53, 6, 0x16],
-  [54, 6, 0x1a],
-  [55, 7, 0x3b],
-  [56, 7, 0x44],
-  [57, 8, 0x91],
-  [58, 8, 0xa5],
-  [59, 8, 0xbe],
-  [60, 9, 0x196],
-  [61, 9, 0x1ae],
-  [62, 9, 0x1b9],
-  [63, 10, 0x3a1],
-  [64, 10, 0x391],
-  [65, 10, 0x3a5],
-  [66, 10, 0x3d5],
-  [67, 8, 0x94],
-  [68, 8, 0x9a],
-  [69, 7, 0x36],
-  [70, 7, 0x38],
-  [71, 7, 0x3a],
-  [72, 7, 0x41],
-  [73, 8, 0x8c],
-  [74, 8, 0x9b],
-  [75, 8, 0xb0],
-  [76, 8, 0xc3],
-  [77, 9, 0x19e],
-  [78, 9, 0x1ab],
-  [79, 9, 0x1bc],
-  [80, 10, 0x39f],
-  [81, 10, 0x38f],
-  [82, 10, 0x3a9],
-  [83, 10, 0x3cf],
-  [84, 8, 0x93],
-  [85, 8, 0xbf],
-  [86, 7, 0x3e],
-  [87, 7, 0x3f],
-  [88, 7, 0x43],
-  [89, 7, 0x45],
-  [90, 8, 0x9e],
-  [91, 8, 0xa7],
-  [92, 8, 0xb9],
-  [93, 9, 0x194],
-  [94, 9, 0x1a2],
-  [95, 9, 0x1ba],
-  [96, 9, 0x1c3],
-  [97, 10, 0x3a6],
-  [98, 10, 0x3a7],
-  [99, 10, 0x3bb],
-  [100, 10, 0x3d4],
-  [101, 8, 0x9f],
-  [102, 9, 0x1a0],
-  [103, 8, 0x8f],
-  [104, 8, 0x8d],
-  [105, 8, 0x90],
-  [106, 8, 0x98],
-  [107, 8, 0xa6],
-  [108, 8, 0xb6],
-  [109, 8, 0xc4],
-  [110, 9, 0x19f],
-  [111, 9, 0x1af],
-  [112, 9, 0x1bf],
-  [113, 10, 0x399],
-  [114, 10, 0x3bf],
-  [115, 10, 0x3b4],
-  [116, 10, 0x3c9],
-  [117, 10, 0x3e7],
-  [118, 8, 0xa8],
-  [119, 9, 0x1b6],
-  [120, 8, 0xab],
-  [121, 8, 0xa4],
-  [122, 8, 0xaa],
-  [123, 8, 0xb2],
-  [124, 8, 0xc2],
-  [125, 8, 0xc5],
-  [126, 9, 0x198],
-  [127, 9, 0x1a4],
-  [128, 9, 0x1b8],
-  [129, 10, 0x38c],
-  [130, 10, 0x3a4],
-  [131, 10, 0x3c4],
-  [132, 10, 0x3c6],
-  [133, 10, 0x3dd],
-  [134, 10, 0x3e8],
-  [135, 8, 0xad],
-  [136, 10, 0x3af],
-  [137, 9, 0x192],
-  [138, 8, 0xbd],
-  [139, 8, 0xbc],
-  [140, 9, 0x18e],
-  [141, 9, 0x197],
-  [142, 9, 0x19a],
-  [143, 9, 0x1a3],
-  [144, 9, 0x1b1],
-  [145, 10, 0x38d],
-  [146, 10, 0x398],
-  [147, 10, 0x3b7],
-  [148, 10, 0x3d3],
-  [149, 10, 0x3d1],
-  [150, 10, 0x3db],
-  [151, 11, 0x7dd],
-  [152, 8, 0xb4],
-  [153, 10, 0x3de],
-  [154, 9, 0x1a9],
-  [155, 9, 0x19b],
-  [156, 9, 0x19c],
-  [157, 9, 0x1a1],
-  [158, 9, 0x1aa],
-  [159, 9, 0x1ad],
-  [160, 9, 0x1b3],
-  [161, 10, 0x38b],
-  [162, 10, 0x3b2],
-  [163, 10, 0x3b8],
-  [164, 10, 0x3ce],
-  [165, 10, 0x3e1],
-  [166, 10, 0x3e0],
-  [167, 11, 0x7d2],
-  [168, 11, 0x7e5],
-  [169, 8, 0xb7],
-  [170, 11, 0x7e3],
-  [171, 9, 0x1bb],
-  [172, 9, 0x1a8],
-  [173, 9, 0x1a6],
-  [174, 9, 0x1b0],
-  [175, 9, 0x1b2],
-  [176, 9, 0x1b7],
-  [177, 10, 0x39b],
-  [178, 10, 0x39a],
-  [179, 10, 0x3ba],
-  [180, 10, 0x3b5],
-  [181, 10, 0x3d6],
-  [182, 11, 0x7d7],
-  [183, 10, 0x3e4],
-  [184, 11, 0x7d8],
-  [185, 11, 0x7ea],
-  [186, 8, 0xba],
-  [187, 11, 0x7e8],
-  [188, 10, 0x3a0],
-  [189, 9, 0x1bd],
-  [190, 9, 0x1b4],
-  [191, 10, 0x38a],
-  [192, 9, 0x1c4],
-  [193, 10, 0x392],
-  [194, 10, 0x3aa],
-  [195, 10, 0x3b0],
-  [196, 10, 0x3bc],
-  [197, 10, 0x3d7],
-  [198, 11, 0x7d4],
-  [199, 11, 0x7dc],
-  [200, 11, 0x7db],
-  [201, 11, 0x7d5],
-  [202, 11, 0x7f0],
-  [203, 8, 0xc1],
-  [204, 11, 0x7fb],
-  [205, 10, 0x3c8],
-  [206, 10, 0x3a3],
-  [207, 10, 0x395],
-  [208, 10, 0x39d],
-  [209, 10, 0x3ac],
-  [210, 10, 0x3ae],
-  [211, 10, 0x3c5],
-  [212, 10, 0x3d8],
-  [213, 10, 0x3e2],
-  [214, 10, 0x3e6],
-  [215, 11, 0x7e4],
-  [216, 11, 0x7e7],
-  [217, 11, 0x7e0],
-  [218, 11, 0x7e9],
-  [219, 11, 0x7f7],
-  [220, 9, 0x190],
-  [221, 11, 0x7f2],
-  [222, 10, 0x393],
-  [223, 9, 0x1be],
-  [224, 9, 0x1c0],
-  [225, 10, 0x394],
-  [226, 10, 0x397],
-  [227, 10, 0x3ad],
-  [228, 10, 0x3c3],
-  [229, 10, 0x3c1],
-  [230, 10, 0x3d2],
-  [231, 11, 0x7da],
-  [232, 11, 0x7d9],
-  [233, 11, 0x7df],
-  [234, 11, 0x7eb],
-  [235, 11, 0x7f4],
-  [236, 11, 0x7fa],
-  [237, 9, 0x195],
-  [238, 11, 0x7f8],
-  [239, 10, 0x3bd],
-  [240, 10, 0x39c],
-  [241, 10, 0x3ab],
-  [242, 10, 0x3a8],
-  [243, 10, 0x3b3],
-  [244, 10, 0x3b9],
-  [245, 10, 0x3d0],
-  [246, 10, 0x3e3],
-  [247, 10, 0x3e5],
-  [248, 11, 0x7e2],
-  [249, 11, 0x7de],
-  [250, 11, 0x7ed],
-  [251, 11, 0x7f1],
-  [252, 11, 0x7f9],
-  [253, 11, 0x7fc],
-  [254, 9, 0x193],
-  [255, 12, 0xffd],
-  [256, 10, 0x3dc],
-  [257, 10, 0x3b6],
-  [258, 10, 0x3c7],
-  [259, 10, 0x3cc],
-  [260, 10, 0x3cb],
-  [261, 10, 0x3d9],
-  [262, 10, 0x3da],
-  [263, 11, 0x7d3],
-  [264, 11, 0x7e1],
-  [265, 11, 0x7ee],
-  [266, 11, 0x7ef],
-  [267, 11, 0x7f5],
-  [268, 11, 0x7f6],
-  [269, 12, 0xffc],
-  [270, 12, 0xfff],
-  [271, 9, 0x19d],
-  [272, 9, 0x1c2],
-  [273, 8, 0xb5],
-  [274, 8, 0xa1],
-  [275, 8, 0x96],
-  [276, 8, 0x97],
-  [277, 8, 0x95],
-  [278, 8, 0x99],
-  [279, 8, 0xa0],
-  [280, 8, 0xa2],
-  [281, 8, 0xac],
-  [282, 8, 0xa9],
-  [283, 8, 0xb1],
-  [284, 8, 0xb3],
-  [285, 8, 0xbb],
-  [286, 8, 0xc0],
-  [287, 9, 0x18f],
-  [288, 5, 0x4]]
+export const scaleFactorCB = [
+  {
+    "0": 60,
+    "size": 1
+  },
+  {
+    "100": 59,
+    "size": 3
+  },
+  {
+    "1010": 61,
+    "1011": 58,
+    "1100": 62,
+    "size": 4
+  },
+  {
+    "11010": 57,
+    "11011": 63,
+    "size": 5
+  },
+  {
+    "111000": 56,
+    "111001": 64,
+    "111010": 55,
+    "111011": 65,
+    "size": 6
+  },
+  {
+    "1111000": 66,
+    "1111001": 54,
+    "1111010": 67,
+    "size": 7
+  },
+  {
+    "11110110": 53,
+    "11110111": 68,
+    "11111000": 52,
+    "11111001": 69,
+    "11111010": 51,
+    "size": 8
+  },
+  {
+    "111110110": 70,
+    "111110111": 50,
+    "111111000": 49,
+    "111111001": 71,
+    "size": 9
+  },
+  {
+    "1111110100": 72,
+    "1111110101": 48,
+    "1111110110": 73,
+    "1111110111": 47,
+    "1111111000": 74,
+    "1111111001": 46,
+    "size": 10
+  },
+  {
+    "size": 11,
+    "11111110101": 75,
+    "11111110100": 76,
+    "11111111001": 43,
+    "11111111000": 45,
+    "11111110110": 77,
+    "11111110111": 78
+  },
+  {
+    "size": 12,
+    "111111110100": 44,
+    "111111110111": 41,
+    "111111110110": 42,
+    "111111111001": 40,
+    "111111110101": 79,
+    "111111111000": 80
+  },
+  {
+    "size": 13,
+    "1111111110100": 81,
+    "1111111110110": 82,
+    "1111111110111": 38,
+    "1111111110101": 39,
+    "1111111111000": 83
+  },
+  {
+    "size": 14,
+    "11111111110010": 37,
+    "11111111110111": 34,
+    "11111111110110": 36,
+    "11111111110101": 33,
+    "11111111111001": 32,
+    "11111111111000": 84,
+    "11111111110100": 85,
+    "11111111110011": 35
+  },
+  {
+    "size": 15,
+    "111111111110110": 30,
+    "111111111110100": 87,
+    "111111111110111": 31,
+    "111111111110101": 89
+  },
+  {
+    "size": 16,
+    "1111111111110011": 27,
+    "1111111111110100": 28,
+    "1111111111110001": 29,
+    "1111111111110010": 26,
+    "1111111111110101": 24,
+    "1111111111110110": 88,
+    "1111111111110000": 86
+  },
+  {
+    "size": 17,
+    "11111111111110000": 23,
+    "11111111111101110": 25,
+    "11111111111101111": 22
+  },
+  {
+    "size": 18,
+    "111111111111100111": 2,
+    "111111111111100100": 19,
+    "111111111111100101": 3,
+    "111111111111100011": 21,
+    "111111111111100010": 90,
+    "111111111111101000": 0,
+    "111111111111100110": 1
+  },
+  {
+    "size": 19,
+    "1111111111111010010": 98,
+    "1111111111111111010": 20,
+    "1111111111111111011": 17,
+    "1111111111111111000": 16,
+    "1111111111111110111": 15,
+    "1111111111111111110": 14,
+    "1111111111111111111": 13,
+    "1111111111111111101": 12,
+    "1111111111111111100": 11,
+    "1111111111111110000": 10,
+    "1111111111111101111": 9,
+    "1111111111111101110": 8,
+    "1111111111111110110": 7,
+    "1111111111111101101": 6,
+    "1111111111111110001": 5,
+    "1111111111111110101": 4,
+    "1111111111111011001": 91,
+    "1111111111111011010": 92,
+    "1111111111111011011": 93,
+    "1111111111111011100": 94,
+    "1111111111111011101": 95,
+    "1111111111111011110": 96,
+    "1111111111111011000": 97,
+    "1111111111111111001": 18,
+    "1111111111111010011": 99,
+    "1111111111111010100": 100,
+    "1111111111111010101": 101,
+    "1111111111111010110": 102,
+    "1111111111111110010": 103,
+    "1111111111111011111": 104,
+    "1111111111111100111": 105,
+    "1111111111111101000": 106,
+    "1111111111111101001": 107,
+    "1111111111111101010": 108,
+    "1111111111111101011": 109,
+    "1111111111111100110": 110,
+    "1111111111111100000": 111,
+    "1111111111111100001": 112,
+    "1111111111111100010": 113,
+    "1111111111111100011": 114,
+    "1111111111111100100": 115,
+    "1111111111111100101": 116,
+    "1111111111111010111": 117,
+    "1111111111111101100": 118,
+    "1111111111111110100": 119,
+    "1111111111111110011": 120
+  }
+];
+
+export const spectrumCB = [
+  [],
+  [
+    {
+      "0": 40,
+      "size": 1
+    },
+    {
+      "10000": 67,
+      "10001": 13,
+      "10010": 39,
+      "10011": 49,
+      "10100": 41,
+      "10101": 37,
+      "10110": 43,
+      "10111": 31,
+      "size": 5
+    },
+    {
+      "1100000": 58,
+      "1100001": 22,
+      "1100010": 38,
+      "1100011": 46,
+      "1100100": 34,
+      "1100101": 42,
+      "1100110": 76,
+      "1100111": 36,
+      "1101000": 4,
+      "1101001": 28,
+      "1101010": 64,
+      "1101011": 48,
+      "1101100": 16,
+      "1101101": 44,
+      "1101110": 70,
+      "1101111": 32,
+      "1110000": 52,
+      "1110001": 50,
+      "1110010": 10,
+      "1110011": 68,
+      "1110100": 12,
+      "1110101": 66,
+      "1110110": 14,
+      "1110111": 30,
+      "size": 7
+    },
+    {
+      "111100000": 73,
+      "111100001": 19,
+      "111100010": 61,
+      "111100011": 51,
+      "111100100": 47,
+      "111100101": 35,
+      "111100110": 33,
+      "111100111": 55,
+      "111101000": 65,
+      "111101001": 45,
+      "111101010": 25,
+      "111101011": 15,
+      "111101100": 7,
+      "111101101": 29,
+      "111101110": 59,
+      "111101111": 57,
+      "111110000": 21,
+      "111110001": 1,
+      "111110010": 27,
+      "111110011": 53,
+      "111110100": 69,
+      "111110101": 77,
+      "111110110": 23,
+      "111110111": 79,
+      "size": 9
+    },
+    {
+      "1111110000": 5,
+      "1111110001": 9,
+      "1111110010": 75,
+      "1111110011": 63,
+      "1111110100": 11,
+      "1111110101": 3,
+      "1111110110": 17,
+      "1111110111": 71,
+      "size": 10
+    },
+    {
+      "size": 11,
+      "11111111101": 2,
+      "11111111110": 54,
+      "11111110101": 8,
+      "11111110011": 56,
+      "11111111000": 0,
+      "11111110010": 24,
+      "11111111100": 18,
+      "11111110110": 72,
+      "11111110000": 60,
+      "11111111001": 74,
+      "11111111011": 26,
+      "11111111010": 62,
+      "11111110111": 6,
+      "11111111111": 78,
+      "11111110001": 20,
+      "11111110100": 80
+    }
+  ],
+  [
+    {
+      "size": 3,
+      "000": 40
+    },
+    {
+      "size": 4,
+      "0010": 67
+    },
+    {
+      "size": 5,
+      "00110": 13,
+      "01010": 31,
+      "01000": 37,
+      "01001": 39,
+      "00111": 41,
+      "01011": 43,
+      "01100": 49
+    },
+    {
+      "100000": 12,
+      "100001": 58,
+      "100010": 64,
+      "100011": 4,
+      "100100": 36,
+      "100101": 70,
+      "100110": 68,
+      "100111": 32,
+      "101000": 16,
+      "101001": 50,
+      "101010": 28,
+      "101011": 14,
+      "101100": 30,
+      "101101": 10,
+      "101110": 76,
+      "101111": 52,
+      "110000": 44,
+      "110001": 66,
+      "size": 6,
+      "011011": 22,
+      "011010": 34,
+      "011111": 38,
+      "011101": 42,
+      "011100": 46,
+      "011110": 48
+    },
+    {
+      "1100100": 47,
+      "1100101": 65,
+      "1100110": 19,
+      "1100111": 33,
+      "1101000": 61,
+      "1101001": 75,
+      "1101010": 71,
+      "1101011": 25,
+      "1101100": 29,
+      "1101101": 79,
+      "1101110": 15,
+      "1101111": 1,
+      "1110000": 11,
+      "1110001": 55,
+      "1110010": 73,
+      "size": 7
+    },
+    {
+      "11100110": 59,
+      "11100111": 21,
+      "11101000": 7,
+      "11101001": 17,
+      "11101010": 5,
+      "11101011": 3,
+      "11101100": 27,
+      "11101101": 69,
+      "11101110": 63,
+      "11101111": 45,
+      "11110000": 53,
+      "11110001": 23,
+      "11110010": 9,
+      "11110011": 51,
+      "11110100": 57,
+      "11110101": 35,
+      "11110110": 77,
+      "11110111": 60,
+      "11111000": 20,
+      "size": 8
+    },
+    {
+      "111110010": 56,
+      "111110011": 0,
+      "111110100": 24,
+      "111110101": 26,
+      "111110110": 80,
+      "111110111": 6,
+      "111111000": 62,
+      "111111001": 18,
+      "111111010": 8,
+      "111111011": 72,
+      "111111100": 54,
+      "111111101": 2,
+      "111111110": 74,
+      "111111111": 78,
+      "size": 9
+    }
+  ],
+  [
+    {
+      "0": 0,
+      "size": 1
+    },
+    {
+      "1000": 27,
+      "1001": 1,
+      "1010": 9,
+      "1011": 3,
+      "size": 4
+    },
+    {
+      "11000": 36,
+      "11001": 4,
+      "size": 5
+    },
+    {
+      "110100": 12,
+      "110101": 10,
+      "110110": 30,
+      "110111": 13,
+      "111000": 28,
+      "111001": 39,
+      "size": 6
+    },
+    {
+      "1110100": 40,
+      "1110101": 31,
+      "1110110": 37,
+      "size": 7
+    },
+    {
+      "11101110": 54,
+      "11101111": 2,
+      "11110000": 5,
+      "11110001": 63,
+      "11110010": 48,
+      "size": 8
+    },
+    {
+      "111100110": 7,
+      "111100111": 16,
+      "111101000": 45,
+      "111101001": 14,
+      "111101010": 66,
+      "111101011": 6,
+      "111101100": 21,
+      "111101101": 15,
+      "111101110": 18,
+      "111101111": 11,
+      "111110000": 57,
+      "111110001": 49,
+      "111110010": 22,
+      "111110011": 42,
+      "111110100": 43,
+      "size": 9
+    },
+    {
+      "1111101010": 46,
+      "1111101011": 33,
+      "1111101100": 34,
+      "1111101101": 19,
+      "1111101110": 67,
+      "1111101111": 41,
+      "1111110000": 64,
+      "1111110001": 32,
+      "1111110010": 8,
+      "1111110011": 17,
+      "1111110100": 75,
+      "1111110101": 51,
+      "1111110110": 29,
+      "1111110111": 55,
+      "1111111000": 25,
+      "size": 10
+    },
+    {
+      "size": 11,
+      "11111111001": 23,
+      "11111111000": 24,
+      "11111110111": 76,
+      "11111110011": 52,
+      "11111110010": 72,
+      "11111110100": 38,
+      "11111110110": 44,
+      "11111110101": 58
+    },
+    {
+      "size": 12,
+      "111111111100": 53,
+      "111111110100": 35,
+      "111111110101": 73,
+      "111111111011": 50,
+      "111111111010": 70,
+      "111111110111": 78,
+      "111111110110": 69,
+      "111111111001": 79,
+      "111111111000": 26
+    },
+    {
+      "size": 13,
+      "1111111111010": 20,
+      "1111111111100": 47,
+      "1111111111011": 60
+    },
+    {
+      "size": 14,
+      "11111111111011": 68,
+      "11111111111010": 61,
+      "11111111111100": 65
+    },
+    {
+      "size": 15,
+      "111111111111010": 80,
+      "111111111111100": 71,
+      "111111111111101": 59,
+      "111111111111011": 77,
+      "111111111111110": 56
+    },
+    {
+      "size": 16,
+      "1111111111111111": 62,
+      "1111111111111110": 74
+    }
+  ],
+  [
+    {
+      "1000": 4,
+      "1001": 30,
+      "size": 4,
+      "0000": 40,
+      "0111": 0,
+      "0011": 39,
+      "0010": 37,
+      "0110": 36,
+      "0100": 31,
+      "0101": 27,
+      "0001": 13
+    },
+    {
+      "10100": 28,
+      "10101": 12,
+      "10110": 1,
+      "10111": 10,
+      "11000": 3,
+      "11001": 9,
+      "size": 5
+    },
+    {
+      "1101000": 67,
+      "1101001": 43,
+      "1101010": 49,
+      "1101011": 41,
+      "1101100": 66,
+      "1101101": 64,
+      "1101110": 48,
+      "1101111": 58,
+      "1110000": 16,
+      "size": 7
+    },
+    {
+      "11100010": 14,
+      "11100011": 42,
+      "11100100": 22,
+      "11100101": 32,
+      "11100110": 46,
+      "11100111": 38,
+      "11101000": 34,
+      "11101001": 63,
+      "11101010": 57,
+      "11101011": 45,
+      "11101100": 55,
+      "11101101": 11,
+      "11101110": 21,
+      "11101111": 5,
+      "11110000": 15,
+      "11110001": 19,
+      "11110010": 29,
+      "11110011": 7,
+      "11110100": 33,
+      "11110101": 54,
+      "11110110": 2,
+      "size": 8
+    },
+    {
+      "111101110": 18,
+      "111101111": 6,
+      "111110000": 52,
+      "111110001": 76,
+      "111110010": 70,
+      "111110011": 44,
+      "111110100": 50,
+      "111110101": 68,
+      "size": 9
+    },
+    {
+      "1111101100": 51,
+      "1111101101": 75,
+      "1111101110": 69,
+      "1111101111": 25,
+      "1111110000": 17,
+      "1111110001": 73,
+      "1111110010": 23,
+      "1111110011": 61,
+      "1111110100": 35,
+      "1111110101": 79,
+      "1111110110": 47,
+      "1111110111": 59,
+      "1111111000": 65,
+      "1111111001": 53,
+      "size": 10
+    },
+    {
+      "size": 11,
+      "11111111010": 20,
+      "11111111100": 80,
+      "11111111001": 60,
+      "11111110100": 71,
+      "11111110111": 72,
+      "11111110110": 24,
+      "11111111000": 8,
+      "11111111011": 56,
+      "11111111101": 26,
+      "11111110101": 77,
+      "11111111110": 78
+    },
+    {
+      "size": 12,
+      "111111111111": 62,
+      "111111111110": 74
+    }
+  ],
+  [
+    {
+      "0": 40,
+      "size": 1
+    },
+    {
+      "1000": 31,
+      "1001": 49,
+      "1010": 41,
+      "1011": 39,
+      "size": 4
+    },
+    {
+      "11000": 48,
+      "11001": 32,
+      "11010": 30,
+      "11011": 50,
+      "size": 5
+    },
+    {
+      "1110000": 22,
+      "1110001": 42,
+      "1110010": 58,
+      "1110011": 38,
+      "size": 7
+    },
+    {
+      "11101000": 21,
+      "11101001": 59,
+      "11101010": 29,
+      "11101011": 51,
+      "11101100": 23,
+      "11101101": 57,
+      "11101110": 33,
+      "11101111": 47,
+      "11110000": 13,
+      "11110001": 67,
+      "11110010": 37,
+      "11110011": 43,
+      "size": 8
+    },
+    {
+      "111101000": 12,
+      "111101001": 52,
+      "111101010": 68,
+      "111101011": 28,
+      "111101100": 14,
+      "111101101": 66,
+      "111101110": 46,
+      "111101111": 34,
+      "111110000": 24,
+      "111110001": 60,
+      "111110010": 20,
+      "111110011": 56,
+      "size": 9
+    },
+    {
+      "1111101000": 11,
+      "1111101001": 65,
+      "1111101010": 25,
+      "1111101011": 55,
+      "1111101100": 69,
+      "1111101101": 61,
+      "1111101110": 15,
+      "1111101111": 19,
+      "1111110000": 36,
+      "1111110001": 4,
+      "1111110010": 77,
+      "1111110011": 76,
+      "size": 10
+    },
+    {
+      "size": 11,
+      "11111101011": 27,
+      "11111101001": 44,
+      "11111101111": 45,
+      "11111101101": 35,
+      "11111110001": 10,
+      "11111110101": 78,
+      "11111110111": 62,
+      "11111110000": 64,
+      "11111110010": 16,
+      "11111101110": 5,
+      "11111101000": 3,
+      "11111111001": 6,
+      "11111101100": 53,
+      "11111111000": 70,
+      "11111101010": 75,
+      "11111110110": 54,
+      "11111110011": 26,
+      "11111110100": 2
+    },
+    {
+      "size": 12,
+      "111111110110": 63,
+      "111111110100": 18,
+      "111111111001": 71,
+      "111111111100": 73,
+      "111111110101": 74,
+      "111111111000": 7,
+      "111111111101": 9,
+      "111111111011": 79,
+      "111111111010": 17,
+      "111111110111": 1
+    },
+    {
+      "size": 13,
+      "1111111111111": 0,
+      "1111111111100": 72,
+      "1111111111101": 8,
+      "1111111111110": 80
+    }
+  ],
+  [
+    {
+      "1000": 30,
+      "size": 4,
+      "0011": 41,
+      "0000": 40,
+      "0101": 50,
+      "0001": 49,
+      "0111": 48,
+      "0010": 39,
+      "0110": 32,
+      "0100": 31
+    },
+    {
+      "100100": 57,
+      "100101": 59,
+      "100110": 23,
+      "100111": 21,
+      "101000": 22,
+      "101001": 33,
+      "101010": 58,
+      "101011": 47,
+      "101100": 51,
+      "101101": 38,
+      "101110": 29,
+      "101111": 42,
+      "110000": 56,
+      "110001": 24,
+      "110010": 20,
+      "110011": 60,
+      "size": 6
+    },
+    {
+      "1101000": 14,
+      "1101001": 68,
+      "1101010": 66,
+      "1101011": 34,
+      "1101100": 12,
+      "1101101": 52,
+      "1101110": 46,
+      "1101111": 28,
+      "1110000": 67,
+      "1110001": 13,
+      "1110010": 37,
+      "1110011": 43,
+      "1110100": 69,
+      "size": 7
+    },
+    {
+      "11101010": 11,
+      "11101011": 25,
+      "11101100": 61,
+      "11101101": 65,
+      "11101110": 55,
+      "11101111": 19,
+      "11110000": 15,
+      "11110001": 70,
+      "size": 8
+    },
+    {
+      "111100100": 64,
+      "111100101": 10,
+      "111100110": 16,
+      "111100111": 45,
+      "111101000": 27,
+      "111101001": 77,
+      "111101010": 5,
+      "111101011": 3,
+      "111101100": 53,
+      "111101101": 75,
+      "111101110": 35,
+      "111101111": 36,
+      "111110000": 6,
+      "111110001": 2,
+      "111110010": 62,
+      "111110011": 18,
+      "111110100": 4,
+      "111110101": 78,
+      "111110110": 74,
+      "111110111": 26,
+      "111111000": 76,
+      "111111001": 54,
+      "111111010": 44,
+      "size": 9
+    },
+    {
+      "1111110110": 9,
+      "1111110111": 17,
+      "1111111000": 63,
+      "1111111001": 73,
+      "1111111010": 71,
+      "1111111011": 79,
+      "1111111100": 7,
+      "1111111101": 1,
+      "size": 10
+    },
+    {
+      "size": 11,
+      "11111111110": 0,
+      "11111111111": 72,
+      "11111111101": 8,
+      "11111111100": 80
+    }
+  ],
+  [
+    {
+      "0": 0,
+      "size": 1
+    },
+    {
+      "100": 8,
+      "101": 1,
+      "size": 3
+    },
+    {
+      "1100": 9,
+      "size": 4
+    },
+    {
+      "110100": 17,
+      "110101": 10,
+      "110110": 16,
+      "110111": 2,
+      "size": 6
+    },
+    {
+      "1110000": 25,
+      "1110001": 11,
+      "1110010": 18,
+      "1110011": 24,
+      "1110100": 3,
+      "size": 7
+    },
+    {
+      "11101010": 19,
+      "11101011": 26,
+      "11101100": 12,
+      "11101101": 33,
+      "11101110": 13,
+      "11101111": 41,
+      "11110000": 27,
+      "11110001": 20,
+      "11110010": 4,
+      "11110011": 32,
+      "size": 8
+    },
+    {
+      "111101000": 34,
+      "111101001": 21,
+      "111101010": 42,
+      "111101011": 5,
+      "111101100": 49,
+      "111101101": 40,
+      "111101110": 14,
+      "111101111": 35,
+      "111110000": 29,
+      "111110001": 28,
+      "111110010": 43,
+      "111110011": 22,
+      "111110100": 50,
+      "111110101": 15,
+      "size": 9
+    },
+    {
+      "1111101100": 30,
+      "1111101101": 6,
+      "1111101110": 48,
+      "1111101111": 36,
+      "1111110000": 57,
+      "1111110001": 37,
+      "1111110010": 58,
+      "1111110011": 44,
+      "1111110100": 51,
+      "1111110101": 23,
+      "1111110110": 59,
+      "1111110111": 52,
+      "1111111000": 45,
+      "1111111001": 38,
+      "1111111010": 31,
+      "size": 10
+    },
+    {
+      "size": 11,
+      "11111111011": 39,
+      "11111111100": 47,
+      "11111111101": 61,
+      "11111111010": 60,
+      "11111110110": 56,
+      "11111111001": 46,
+      "11111111000": 53,
+      "11111110111": 7
+    },
+    {
+      "size": 12,
+      "111111111110": 55,
+      "111111111101": 54,
+      "111111111100": 62,
+      "111111111111": 63
+    }
+  ],
+  [
+    {
+      "size": 3,
+      "000": 9
+    },
+    {
+      "size": 4,
+      "0101": 1,
+      "0011": 8,
+      "0100": 10,
+      "0010": 17,
+      "0110": 18
+    },
+    {
+      "10000": 2,
+      "10001": 25,
+      "10010": 11,
+      "10011": 26,
+      "10100": 19,
+      "size": 5,
+      "01111": 16,
+      "01110": 0
+    },
+    {
+      "101010": 27,
+      "101011": 33,
+      "101100": 12,
+      "101101": 34,
+      "101110": 20,
+      "101111": 24,
+      "110000": 3,
+      "110001": 35,
+      "110010": 28,
+      "110011": 42,
+      "size": 6
+    },
+    {
+      "1101000": 41,
+      "1101001": 21,
+      "1101010": 13,
+      "1101011": 43,
+      "1101100": 29,
+      "1101101": 36,
+      "1101110": 44,
+      "1101111": 4,
+      "1110000": 37,
+      "1110001": 32,
+      "1110010": 22,
+      "1110011": 50,
+      "1110100": 49,
+      "1110101": 14,
+      "size": 7
+    },
+    {
+      "11101100": 30,
+      "11101101": 51,
+      "11101110": 45,
+      "11101111": 40,
+      "11110000": 52,
+      "11110001": 5,
+      "11110010": 38,
+      "11110011": 57,
+      "11110100": 58,
+      "11110101": 23,
+      "11110110": 53,
+      "11110111": 59,
+      "11111000": 15,
+      "11111001": 46,
+      "11111010": 31,
+      "size": 8
+    },
+    {
+      "111110110": 54,
+      "111110111": 60,
+      "111111000": 48,
+      "111111001": 39,
+      "111111010": 6,
+      "111111011": 61,
+      "111111100": 62,
+      "111111101": 55,
+      "size": 9
+    },
+    {
+      "1111111100": 47,
+      "1111111101": 56,
+      "1111111110": 7,
+      "1111111111": 63,
+      "size": 10
+    }
+  ],
+  [
+    {
+      "0": 0,
+      "size": 1
+    },
+    {
+      "100": 13,
+      "101": 1,
+      "size": 3
+    },
+    {
+      "1100": 14,
+      "size": 4
+    },
+    {
+      "110100": 27,
+      "110101": 15,
+      "110110": 26,
+      "110111": 2,
+      "size": 6
+    },
+    {
+      "1110000": 40,
+      "1110001": 28,
+      "1110010": 16,
+      "size": 7
+    },
+    {
+      "11100110": 39,
+      "11100111": 3,
+      "11101000": 29,
+      "11101001": 41,
+      "11101010": 17,
+      "11101011": 53,
+      "11101100": 30,
+      "11101101": 18,
+      "size": 8
+    },
+    {
+      "111011100": 54,
+      "111011101": 42,
+      "111011110": 4,
+      "111011111": 52,
+      "111100000": 66,
+      "111100001": 31,
+      "111100010": 19,
+      "111100011": 43,
+      "111100100": 67,
+      "111100101": 79,
+      "111100110": 55,
+      "size": 9
+    },
+    {
+      "1111001110": 5,
+      "1111001111": 32,
+      "1111010000": 65,
+      "1111010001": 20,
+      "1111010010": 44,
+      "1111010011": 21,
+      "1111010100": 105,
+      "1111010101": 56,
+      "1111010110": 68,
+      "1111010111": 80,
+      "1111011000": 92,
+      "1111011001": 6,
+      "1111011010": 106,
+      "1111011011": 34,
+      "1111011100": 45,
+      "1111011101": 33,
+      "1111011110": 57,
+      "1111011111": 118,
+      "1111100000": 22,
+      "1111100001": 93,
+      "size": 10
+    },
+    {
+      "size": 11,
+      "11111001100": 46,
+      "11111011100": 60,
+      "11111000101": 69,
+      "11111010001": 70,
+      "11111011011": 71,
+      "11111100000": 73,
+      "11111000100": 78,
+      "11111001011": 58,
+      "11111011110": 48,
+      "11111000110": 81,
+      "11111001111": 82,
+      "11111011010": 83,
+      "11111001010": 47,
+      "11111011111": 144,
+      "11111010011": 91,
+      "11111011101": 59,
+      "11111001000": 7,
+      "11111010100": 94,
+      "11111011001": 95,
+      "11111010101": 132,
+      "11111001110": 131,
+      "11111010010": 104,
+      "11111001101": 8,
+      "11111010000": 35,
+      "11111000111": 107,
+      "11111010111": 108,
+      "11111100010": 109,
+      "11111010110": 120,
+      "11111001001": 119,
+      "11111100001": 117,
+      "11111011000": 23
+    },
+    {
+      "size": 12,
+      "111111000111": 36,
+      "111111010100": 37,
+      "111111001101": 61,
+      "111111100010": 62,
+      "111111011010": 85,
+      "111111100011": 86,
+      "111111101001": 87,
+      "111111100111": 63,
+      "111111101000": 143,
+      "111111100001": 135,
+      "111111011100": 147,
+      "111111100100": 38,
+      "111111011000": 49,
+      "111111101010": 50,
+      "111111001100": 157,
+      "111111010011": 96,
+      "111111011110": 97,
+      "111111010001": 134,
+      "111111000110": 133,
+      "111111010110": 158,
+      "111111001011": 84,
+      "111111100000": 130,
+      "111111100110": 124,
+      "111111001111": 24,
+      "111111010010": 72,
+      "111111010101": 25,
+      "111111011001": 74,
+      "111111101011": 75,
+      "111111010111": 146,
+      "111111001110": 110,
+      "111111011011": 111,
+      "111111100101": 123,
+      "111111010000": 122,
+      "111111001010": 121,
+      "111111001001": 145,
+      "111111001000": 9,
+      "111111011101": 10,
+      "111111011111": 159
+    },
+    {
+      "size": 13,
+      "1111111110100": 115,
+      "1111111100110": 88,
+      "1111111011000": 112,
+      "1111111110110": 103,
+      "1111111101011": 125,
+      "1111111101111": 126,
+      "1111111110001": 102,
+      "1111111101010": 101,
+      "1111111100010": 100,
+      "1111111011001": 99,
+      "1111111011101": 98,
+      "1111111110111": 90,
+      "1111111100000": 136,
+      "1111111101000": 137,
+      "1111111110000": 138,
+      "1111111110011": 89,
+      "1111111101110": 113,
+      "1111111101001": 77,
+      "1111111100011": 76,
+      "1111111100001": 64,
+      "1111111011100": 148,
+      "1111111011111": 149,
+      "1111111101101": 150,
+      "1111111110101": 151,
+      "1111111100111": 156,
+      "1111111011011": 51,
+      "1111111101100": 12,
+      "1111111100100": 11,
+      "1111111011110": 160,
+      "1111111011010": 161,
+      "1111111100101": 162,
+      "1111111110010": 163
+    },
+    {
+      "size": 14,
+      "11111111111101": 167,
+      "11111111111001": 152,
+      "11111111111011": 153,
+      "11111111111100": 166,
+      "11111111110111": 165,
+      "11111111110101": 129,
+      "11111111110000": 114,
+      "11111111110010": 116,
+      "11111111110011": 127,
+      "11111111110100": 128,
+      "11111111110001": 139,
+      "11111111111000": 140,
+      "11111111110110": 141,
+      "11111111111010": 164
+    },
+    {
+      "size": 15,
+      "111111111111110": 155,
+      "111111111111101": 154,
+      "111111111111100": 142,
+      "111111111111111": 168
+    }
+  ],
+  [
+    {
+      "size": 4,
+      "0010": 27,
+      "0000": 14,
+      "0001": 15
+    },
+    {
+      "size": 5,
+      "01101": 42,
+      "01001": 16,
+      "01000": 1,
+      "00111": 13,
+      "00110": 28,
+      "01100": 29,
+      "01011": 40,
+      "01010": 41
+    },
+    {
+      "100000": 17,
+      "100001": 53,
+      "100010": 0,
+      "100011": 55,
+      "100100": 43,
+      "100101": 39,
+      "100110": 3,
+      "100111": 56,
+      "101000": 31,
+      "101001": 67,
+      "size": 6,
+      "011110": 30,
+      "011101": 2,
+      "011100": 26,
+      "011111": 54
+    },
+    {
+      "1010100": 18,
+      "1010101": 66,
+      "1010110": 68,
+      "1010111": 44,
+      "1011000": 69,
+      "1011001": 57,
+      "1011010": 80,
+      "1011011": 32,
+      "1011100": 81,
+      "1011101": 52,
+      "1011110": 79,
+      "1011111": 4,
+      "1100000": 19,
+      "1100001": 45,
+      "1100010": 70,
+      "1100011": 82,
+      "1100100": 58,
+      "size": 7
+    },
+    {
+      "11001010": 83,
+      "11001011": 93,
+      "11001100": 46,
+      "11001101": 33,
+      "11001110": 71,
+      "11001111": 106,
+      "11010000": 94,
+      "11010001": 65,
+      "11010010": 92,
+      "11010011": 5,
+      "11010100": 105,
+      "11010101": 20,
+      "11010110": 107,
+      "11010111": 95,
+      "11011000": 59,
+      "11011001": 34,
+      "11011010": 84,
+      "11011011": 96,
+      "11011100": 21,
+      "11011101": 47,
+      "11011110": 108,
+      "11011111": 60,
+      "11100000": 72,
+      "11100001": 109,
+      "11100010": 73,
+      "size": 8
+    },
+    {
+      "111000110": 97,
+      "111000111": 85,
+      "111001000": 119,
+      "111001001": 78,
+      "111001010": 86,
+      "111001011": 120,
+      "111001100": 48,
+      "111001101": 118,
+      "111001110": 35,
+      "111001111": 6,
+      "111010000": 110,
+      "111010001": 121,
+      "111010010": 61,
+      "111010011": 132,
+      "111010100": 22,
+      "111010101": 98,
+      "111010110": 111,
+      "111010111": 122,
+      "111011000": 99,
+      "111011001": 133,
+      "111011010": 74,
+      "111011011": 134,
+      "111011100": 36,
+      "111011101": 131,
+      "111011110": 49,
+      "111011111": 123,
+      "111100000": 87,
+      "111100001": 104,
+      "111100010": 62,
+      "111100011": 91,
+      "111100100": 145,
+      "size": 9
+    },
+    {
+      "1111001010": 100,
+      "1111001011": 146,
+      "1111001100": 136,
+      "1111001101": 23,
+      "1111001110": 144,
+      "1111001111": 124,
+      "1111010000": 7,
+      "1111010001": 112,
+      "1111010010": 135,
+      "1111010011": 50,
+      "1111010100": 75,
+      "1111010101": 113,
+      "1111010110": 148,
+      "1111010111": 8,
+      "1111011000": 147,
+      "1111011001": 37,
+      "1111011010": 101,
+      "1111011011": 88,
+      "1111011100": 137,
+      "1111011101": 63,
+      "1111011110": 24,
+      "1111011111": 158,
+      "1111100000": 125,
+      "1111100001": 159,
+      "1111100010": 149,
+      "1111100011": 76,
+      "1111100100": 160,
+      "1111100101": 150,
+      "1111100110": 161,
+      "1111100111": 51,
+      "1111101000": 89,
+      "1111101001": 117,
+      "1111101010": 138,
+      "1111101011": 130,
+      "1111101100": 157,
+      "1111101101": 9,
+      "1111101110": 64,
+      "1111101111": 126,
+      "1111110000": 162,
+      "1111110001": 38,
+      "1111110010": 114,
+      "size": 10
+    },
+    {
+      "size": 11,
+      "11111110111": 154,
+      "11111101101": 139,
+      "11111111001": 141,
+      "11111101111": 164,
+      "11111110010": 143,
+      "11111101001": 163,
+      "11111111000": 128,
+      "11111100110": 127,
+      "11111111011": 116,
+      "11111101110": 115,
+      "11111110000": 10,
+      "11111110001": 103,
+      "11111101000": 151,
+      "11111110100": 152,
+      "11111110101": 153,
+      "11111110011": 140,
+      "11111101011": 77,
+      "11111111010": 156,
+      "11111101010": 102,
+      "11111110110": 11,
+      "11111100111": 25,
+      "11111101100": 90
+    },
+    {
+      "size": 12,
+      "111111111101": 12,
+      "111111111011": 155,
+      "111111111010": 129,
+      "111111111001": 142,
+      "111111111000": 165,
+      "111111111110": 166,
+      "111111111100": 167,
+      "111111111111": 168
+    }
+  ],
+  [
+    {
+      "size": 4,
+      "0000": 0,
+      "0001": 18
+    },
+    {
+      "size": 5,
+      "00110": 1,
+      "01001": 36,
+      "00111": 35,
+      "01000": 19,
+      "00100": 288,
+      "00101": 17
+    },
+    {
+      "size": 6,
+      "011010": 54,
+      "011001": 2,
+      "010100": 20,
+      "010110": 53,
+      "010111": 34,
+      "010101": 52,
+      "011000": 37
+    },
+    {
+      "1000000": 39,
+      "1000001": 72,
+      "1000010": 22,
+      "1000011": 88,
+      "1000100": 56,
+      "1000101": 89,
+      "size": 7,
+      "0110110": 69,
+      "0111001": 38,
+      "0111100": 51,
+      "0111101": 3,
+      "0110111": 21,
+      "0111011": 55,
+      "0111000": 70,
+      "0111010": 71,
+      "0111110": 86,
+      "0111111": 87
+    },
+    {
+      "10001100": 73,
+      "10001101": 104,
+      "10001110": 40,
+      "10001111": 103,
+      "10010000": 105,
+      "10010001": 57,
+      "10010010": 23,
+      "10010011": 84,
+      "10010100": 67,
+      "10010101": 277,
+      "10010110": 275,
+      "10010111": 276,
+      "10011000": 106,
+      "10011001": 278,
+      "10011010": 68,
+      "10011011": 74,
+      "10011100": 4,
+      "10011101": 50,
+      "10011110": 90,
+      "10011111": 101,
+      "10100000": 279,
+      "10100001": 274,
+      "10100010": 280,
+      "10100011": 41,
+      "10100100": 121,
+      "10100101": 58,
+      "10100110": 107,
+      "10100111": 91,
+      "10101000": 118,
+      "10101001": 282,
+      "10101010": 122,
+      "10101011": 120,
+      "10101100": 281,
+      "10101101": 135,
+      "10101110": 33,
+      "10101111": 24,
+      "10110000": 75,
+      "10110001": 283,
+      "10110010": 123,
+      "10110011": 284,
+      "10110100": 152,
+      "10110101": 273,
+      "10110110": 108,
+      "10110111": 169,
+      "10111000": 42,
+      "10111001": 92,
+      "10111010": 186,
+      "10111011": 285,
+      "10111100": 139,
+      "10111101": 138,
+      "10111110": 59,
+      "10111111": 85,
+      "11000000": 286,
+      "11000001": 203,
+      "11000010": 124,
+      "11000011": 76,
+      "11000100": 109,
+      "11000101": 125,
+      "11000110": 5,
+      "size": 8
+    },
+    {
+      "110001110": 140,
+      "110001111": 287,
+      "110010000": 220,
+      "110010001": 25,
+      "110010010": 137,
+      "110010011": 254,
+      "110010100": 93,
+      "110010101": 237,
+      "110010110": 60,
+      "110010111": 141,
+      "110011000": 126,
+      "110011001": 43,
+      "110011010": 142,
+      "110011011": 155,
+      "110011100": 156,
+      "110011101": 271,
+      "110011110": 77,
+      "110011111": 110,
+      "110100000": 102,
+      "110100001": 157,
+      "110100010": 94,
+      "110100011": 143,
+      "110100100": 127,
+      "110100101": 26,
+      "110100110": 173,
+      "110100111": 6,
+      "110101000": 172,
+      "110101001": 154,
+      "110101010": 158,
+      "110101011": 78,
+      "110101100": 44,
+      "110101101": 159,
+      "110101110": 61,
+      "110101111": 111,
+      "110110000": 174,
+      "110110001": 144,
+      "110110010": 175,
+      "110110011": 160,
+      "110110100": 190,
+      "110110101": 27,
+      "110110110": 119,
+      "110110111": 176,
+      "110111000": 128,
+      "110111001": 62,
+      "110111010": 95,
+      "110111011": 171,
+      "110111100": 79,
+      "110111101": 189,
+      "110111110": 223,
+      "110111111": 112,
+      "111000000": 224,
+      "111000001": 45,
+      "111000010": 272,
+      "111000011": 96,
+      "111000100": 192,
+      "size": 9
+    },
+    {
+      "1110001010": 191,
+      "1110001011": 161,
+      "1110001100": 129,
+      "1110001101": 145,
+      "1110001110": 16,
+      "1110001111": 81,
+      "1110010000": 7,
+      "1110010001": 64,
+      "1110010010": 193,
+      "1110010011": 222,
+      "1110010100": 225,
+      "1110010101": 207,
+      "1110010110": 47,
+      "1110010111": 226,
+      "1110011000": 146,
+      "1110011001": 113,
+      "1110011010": 178,
+      "1110011011": 177,
+      "1110011100": 240,
+      "1110011101": 208,
+      "1110011110": 28,
+      "1110011111": 80,
+      "1110100000": 188,
+      "1110100001": 63,
+      "1110100010": 30,
+      "1110100011": 206,
+      "1110100100": 130,
+      "1110100101": 65,
+      "1110100110": 97,
+      "1110100111": 98,
+      "1110101000": 242,
+      "1110101001": 82,
+      "1110101010": 194,
+      "1110101011": 241,
+      "1110101100": 209,
+      "1110101101": 227,
+      "1110101110": 210,
+      "1110101111": 136,
+      "1110110000": 195,
+      "1110110001": 46,
+      "1110110010": 162,
+      "1110110011": 243,
+      "1110110100": 115,
+      "1110110101": 180,
+      "1110110110": 257,
+      "1110110111": 147,
+      "1110111000": 163,
+      "1110111001": 244,
+      "1110111010": 179,
+      "1110111011": 99,
+      "1110111100": 196,
+      "1110111101": 239,
+      "1110111110": 48,
+      "1110111111": 114,
+      "1111000000": 29,
+      "1111000001": 229,
+      "1111000010": 8,
+      "1111000011": 228,
+      "1111000100": 131,
+      "1111000101": 211,
+      "1111000110": 132,
+      "1111000111": 258,
+      "1111001000": 205,
+      "1111001001": 116,
+      "1111001010": 49,
+      "1111001011": 260,
+      "1111001100": 259,
+      "1111001101": 31,
+      "1111001110": 164,
+      "1111001111": 83,
+      "1111010000": 245,
+      "1111010001": 149,
+      "1111010010": 230,
+      "1111010011": 148,
+      "1111010100": 100,
+      "1111010101": 66,
+      "1111010110": 181,
+      "1111010111": 197,
+      "1111011000": 212,
+      "1111011001": 261,
+      "1111011010": 262,
+      "1111011011": 150,
+      "1111011100": 256,
+      "1111011101": 133,
+      "1111011110": 153,
+      "1111011111": 9,
+      "1111100000": 166,
+      "1111100001": 165,
+      "1111100010": 213,
+      "1111100011": 246,
+      "1111100100": 183,
+      "1111100101": 247,
+      "1111100110": 214,
+      "1111100111": 117,
+      "1111101000": 134,
+      "size": 10
+    },
+    {
+      "size": 11,
+      "11111111000": 238,
+      "11111110100": 235,
+      "11111101011": 234,
+      "11111011111": 233,
+      "11111011001": 232,
+      "11111011010": 231,
+      "11111110010": 221,
+      "11111110111": 219,
+      "11111100010": 248,
+      "11111011110": 249,
+      "11111101101": 250,
+      "11111110001": 251,
+      "11111111001": 252,
+      "11111111100": 253,
+      "11111101001": 218,
+      "11111100110": 10,
+      "11111100000": 217,
+      "11111100111": 216,
+      "11111100100": 215,
+      "11111111011": 204,
+      "11111110000": 202,
+      "11111010101": 201,
+      "11111011011": 200,
+      "11111010011": 263,
+      "11111111010": 236,
+      "11111101110": 265,
+      "11111101111": 266,
+      "11111110101": 267,
+      "11111110110": 268,
+      "11111110011": 11,
+      "11111101100": 13,
+      "11111011100": 199,
+      "11111010100": 198,
+      "11111101000": 187,
+      "11111101010": 185,
+      "11111011000": 184,
+      "11111010111": 182,
+      "11111100011": 170,
+      "11111100101": 168,
+      "11111010010": 167,
+      "11111011101": 151,
+      "11111010110": 32,
+      "11111100001": 264
+    },
+    {
+      "size": 12,
+      "111111111110": 15,
+      "111111111010": 14,
+      "111111111111": 270,
+      "111111111011": 12,
+      "111111111100": 269,
+      "111111111101": 255
+    }
+  ]
 ];
