+++ src/lib/combinators.js
@@ -15,13 +15,21 @@ export const start = function (name, parseFn) {
   return {
     decode: (input, options, output) => {
       let rawBitString = typedArrayToBitString(input);
-      let bitString = removeRBSPTrailingBits(rawBitString);
+      let bitString = rawBitString;
+
+      options = options || {};
+      output = output || {};
+
+      if (!options.no_trailer_bits) {
+        bitString = removeRBSPTrailingBits(rawBitString);
+      }
       let expGolombDecoder = new ExpGolombDecoder(bitString);
 
-      output = output || {};
-      options = options || {};
-
-      return parseFn.decode(expGolombDecoder, output, options);
+      try {
+        return parseFn.decode(expGolombDecoder, output, options);
+      } catch (e) {
+        return e;
+      }
     },
     encode: (input, options) => {
       let expGolombEncoder = new ExpGolombEncoder();
@@ -32,7 +40,6 @@ export const start = function (name, parseFn) {
 
       let output = expGolombEncoder.bitReservoir;
       let bitString = appendRBSPTrailingBits(output);
-      console.log(bitString)
       let data = bitStringToTypedArray(bitString);
 
       return data;
